+++ src/routes/feedRouter.ts
@@ -1,6 +1,9 @@
 import express from "express";
 import { getFeed } from "../controller/getFeed";
+import { FeedController } from "../controller/FeedController"
 
 export const feedRouter = express.Router();
 
-feedRouter.get("/", getFeed);
+const feedController = new FeedController();
+
+feedRouter.get("/", feedController.getFeed);
