+++ src/routes/userRouter.ts
@@ -1,12 +1,9 @@
 import express from "express";
-import { signUp } from "../controller/signUp";
-import { login } from "../controller/login";
-import { makeFriendship } from "../controller/makeFriendship";
-import { undoFriendship } from "../controller/undoFriendship";
+import { UserController } from "../controller/UserController";
 
 export const userRouter = express.Router();
 
-userRouter.post("/signup", signUp);
-userRouter.post("/login", login);
-userRouter.post("/friendship", makeFriendship);
-userRouter.post("/unfollow", undoFriendship);
+const userController = new UserController();
+
+userRouter.post("/signup", userController.signUp);
+userRouter.post("/login", userController.login);
